feat(dracut-install): split ldd command arguments for execvp()
This restores a not-so-subtle feature of previously executing ldd
via popen(), i.e. the ability to use a cross-compiled sysroot.

The ldd command may be passed in via the DRACUT_LDD environment
variable, and the command may contain command line arguments.
The number of such arguments are not known in advance.

Split the command into executable and arguments and run it
via execvp().

Fixes: d010fa0d7f8ef42ad31729d027d2e4be6dd6e588
Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>