fix(shell-completion): remove hashbang from bash completions
Debian's lintian complains about bash-completion-with-hashbang: "This
file starts with the #! sequence that marks interpreted scripts, but it
is a bash completion script that is merely intended to be sourced.
Please remove the line with hashbang, including any designated
interpreter. Please refer to
https://salsa.debian.org/lintian/lintian/-/merge_requests/292#note_139494
for details."

Remove hashbang from bash completions and add a shellcheck directive
instead.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>