fix(base): perform SELinux policy load after switch_root
Perform SELinux policy load after switch_root. This
fixes the bootup process with recent kernels, as it
was getting stuck on Permission Denied errors, due
to the early SELinux policy load.

Thanks to Laszlo Gombos for reviewing this patch and
suggesting to obsolete the SELinux load policy module
instead of removing it completely.

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 modules.d/98selinux/selinux-loadpolicy.sh |    5 +-
 modules.d/99base/init.sh                  |   61 ++++++++++++++++++++++++++++++
 2 files changed, 65 insertions(+), 1 deletion(-)