feat: create a documentation site
This is a proposal to create a documentation site with Antora
(https://antora.org/) which works with the existing asciidoc files.

As motivation; there is a somewhat difficult to navigate (due to length and no
indexing) single .html file made already, and a wiki page that is kept
separately and links back to several docs, which can be viewed via github's
rendering.  I'm of the feeling that incorporating docs into the main workflow
leads to a better outcome, especially taking advantage of pipelines to publish
on commit ensuring things are kept up-to-date.

There isn't much too this other than the framework stuff and some minor syntax
fixes to some of the files.  "make doc_site" outputs the site in build/.

man/dracut.asc becomes the index page; everything else remains in place where
it was.  After discussion we decided to remove the dracut.html file and
generation to avoid confusion over what is current.

I've added in custom css and lightly styled the output site to keep the vibe of
the older html page.

In terms of developer workflow, I know little about nodejs and have not found
it onerous to figure out; "nodejs" installed by your distro and "npx antora"
"just works"; it's very self-contained.

I've included the sample github workflow to publish it to github pages on
commit taken from the Antora docs -- it may need tweaking by project admins.

The output is basic but functional.  The nice thing about a framework like this
is the almost limitless options for improvement :)