feat(lsinitrd): add support for erofs images
Add support to handle erofs images in lsinitrd. Unfortunately the erofs
tooling is missing some functionality of unsquashfs, esp. the ability to
extract single files and list the content of the image. Work around this
deficiency by always extracting the full image and emulate the missing
functionality as close as possible.

While at it also handle the rename of the squashfs image to
squashfs-root.img.

Signed-off-by: Philipp Rudo <prudo@redhat.com>