fix(dracut-install): do not copy SELinux context
SELinux is normally not enabled in the initramfs and thus the file
contexts are not relevant.

On the other hand using the source file context requires dracut-install
to have the setattr and unlink permissions on these source contexts,
e.g. lib_t or modules_object_t, which only very few domains should have
modification rights to, e.g. the system administrator or the package
manager.