test(SYSTEMD-INITRD): kernel-independent systemd-based initrd
Explicitly verify that dracut systemd services are not included in this
test.

Separate systemd-based initrd out from the rest of the test
infrastructure to make it easier to check (e.g. by passing -v)
what is included in this minimal initrd.

This test produces initrd that is comparable design to the systemd
based initrd produced by mkosi or mkinitcpio.

Copy some additional tests steps from FULL-SYSTEMD and include it
in this test as well.