fix(systemd): don't unconditionally install libgcrypt
systemd v256 started loading libgcrypt with dlopen, so this module
started installing it unconditionally. However, the only part of systemd
that uses it without explicitly linking it is the journal forward secure
sealing feature. This module does not install journald or anything else
that is likely to need libgcrypt. Even if some other part did need it,
it would be pulled in using the usual ldd mechanism.

This change also benefits older systemd versions because systemd can be
built without any libgcrypt dependency at all.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>