feat: config example for cloud provider uki vm
This is an example configuration for a 'unified kernel image'.

The key benefit of a unified kernel is that its secure boot
signature covers the initrd and cmdline contents, allowing
a trustworthy measured boot process with attestation, which
is not practical with locally generated initrds/cmdlines.

The initrd in this example only needs a very small set of
block driver modules present, in order to be usable across
KVM, Hyper-V and Xen hypervisors which will cover essentially
all common public and private clouds.