docs: link man pages in doc site
Currently the man pages in the generated doc site don't link to each other.  A
standard link: approach doesn't work that well when the target is both a groff
man page and the HTML site, as you don't want the links in the man page.

This is something other projects have faced, git in particular uses adoc for
man pages and the approach here is styled on its "gitlink:" macro.  The .js
version is for Antora which uses the asciidoctorjs internally and there is an
asciidoc plugin for the man page generation.  The ruby version is included just
in case we want to geneate man pages from asciidoctor directly one day.

For the HTML output it is a link to the page (or, can be marked for external
looukup for other tools).  For the man pages, it's just a usual "strong" text.

I've been through the existing references to convert them to the new macro.