docs: link man pages in doc site
Currently the man pages in the generated doc site don't link to each other.  A
standard link: approach doesn't work that well when the target is both a groff
man page and the HTML site, as you don't want the links in the man page.

This is something other projects have faced, git in particular uses adoc for
man pages and the approach here is styled on its "gitlink:" macro.  The ruby
version is for asciidoctor that creates the man pages; the .js version is for
Antora which uses the asciidoctorjs internally.  For the HTML output it is a
link to the page (or, can be marked for external looukup for other tools).  For
the man pages, it's just a usual "strong" text.

The only sticking point for this is that using these exensions require building
with asciidoctor, rather than the adoc->xml->man path currently being used.
This would be a change for packagers generating the man pages.

I've been through the existing references to convert them to the new macro.