docs: small cleanup of main landing page
This is a few small cleanups.  The message about the license first up makes it
seem like dracut itself is under creative commons license, which isn't true.
I've moved this down to the bottom, and added a note that dracut is gpl2 but
some parts of the docs are CC.

I've added a shorter introduction and de-dented things.  I've also linked in
the development site along with the other man pages.