docs: index page rework
This reworks the index page, which currently has a couple of missing words and
doesn't flow that well.

I've tried to update the references a bit and rework it so that someone with a
basic Linux knowledge but not specifically dracut boot experience could get a
sense of what is going on reading top-to-bottom.