docs: switch to asciidoctor for man generation, with fallback
This switches the man page generator to asciidoctor, with a fallback option to
configure to use the existing asciidoc/docbook path.

The reason to do this is to consolidate the published site generation and man
page generation around asciidoctor.  With that in place, we can use some
asciidoctor features like extensions to better link the man pages together, or
provide other tweaks to the documentation.

This switches the default generation to asciidoctor, but adds a
--disable-asciidoctor switch as a temporary escape hatch for packagers.

The testing is converted to asciidoctor, with an additional test of the
asciidoc toolchain just to avoid major regressions there.

Documentation is updated.

Depends-On: https://github.com/dracut-ng/dracut-ng/pull/783