docs(bash): remove section on optional args
It's unclear under what circumstances these examples are appropriate and they
are somewhat conflicting with standard use of :+/+ in parameter expansions (see
the issue).

Rather than trying to be overally perscriptive about this, just remove it.  We
should make sure to add more explicit reasoning if we add things back in here.

Closes: https://github.com/dracut-ng/dracut-ng/issues/701