fix(dm): remove 59-persistent-storage-dm.rules
59-persistent-storage-dm.rules has been obsolete for a long time.
The blkid builtin is run from 13-dm.rules. The "dmsetup info" call
in this file is actually harmful, because it overrides properties
that have already been set in 10-dm.rules, in particular DM_SUSPENDED,
to which it assigns a string like "Active" or "Suspended", whereas
the stock DM rules and systemd rules expect the property to have
the value "0", or "1", respectively.

Signed-off-by: Martin Wilck <mwilck@suse.com>