test: print all server startup logs in V=1 or V=2 mode
In V=0 mode only print "Waiting for the server to startup" once and do
not print anything from the server startup logs.

In V=1 and V=2 print all lines from the server startup while waiting
for the server to be ready to ease debugging (e.g. when the kernel
crashes the last 10 lines will not be enough).