fix(shell-interpreter): move later in the module ordering
This PR ensures that shell dependencies (e.g. on bash) is
computed before shell-interpreter gets executed.

Do not explicitly depend on shell-interpreter as this would trigger
calling the shell-interpreter depend() function too
early in the model evaluation process.

The impact of this PR can be observed in the BASIC test when
running on void CI containers. Before this PR both 'dash' and
'bash' dracut modules gets included in the initrd.

After the PR only the 'bash' module gets included in the initrd.

Fixes: #860