fix(plymouth): change severity of shutdown log messages
Right now, if someone is using plymouth and the "quiet" boot parameter,
odds are that they do so in the hope of a more polished boot experience.

The shutdown module prints a couple of warnings and then shuts down
plymouth, meaning that these warnings will flash briefly on screen
before the actual reboot/shutdown.

None of the warnings qualify as warnings though:

1. Printing "Unmounted /oldroot" just means that unmounting /oldroot
   (an internal implementational detail) succeeded.
2. Printing "Killing all remaining processes" hardly qualifies as a
   warning since it's a quite expected result of a reboot.

So downgrade them to "info" (which will respect the "quiet" boot
parameter), leading to a nice, quiet reboot.