test(SYSTEMD-INITRD): fix systemd.volatile=overlay testcase
systemd.volatile=overlayfs is invalid value and gets ignored by
systemd. Change it to systemd.volatile=overlay instead.

This test however fails with FULL-SYSTEMD. Remove it for now
as the fix is not obvious, and likely is a systemd bug and
out of scope for the dracut project.

Confirmed that the overlay started as expected

[  OK  ] Mounted /sysroot.
         Starting Enforce Volatile Root File Systems...
[  OK  ] Finished Enforce Volatile Root File Systems.