feat(systemd-sysusers): run systemd-sysusers as part of the build process
This PR makes the boot process faster and the generated initrd smaller.

It also make the code easier to maintain.

The primary goal of this PR is refactoring and removing code,
this is why it is not marked as perf() in the commit.

That said, with this PR systemd-sysusers no longer needs to be copied
into the initrd, so that is about 50KB saving both on when initrd gets
saved/compressed and when it gets loaded and uncompressed.

This PR also make the debugging slightly easier as one can just
inspect /etc/passwd with lsinitrd instead of trying to figure out
which users are created runtime.