ci: demote dmraid testing
dmraid is no longer activly maitaned and has been already removed
form the following distributions
 - CentOS
 - Debian
 - Ubuntu

Dracut will continue to support dmraid, but lets minimize the CI
burdon and only test dmraid on Arch container going forward.

One of the other motivation of this change is to align Fedora
configuration a bit more with CentOS configuration for ease of
maitenance.