test(KERNEL-INSTALL): include generating and testing rescue boot
Make sure kernel-install generates the regular hostonly initrd and
the rescue (non-hostonly) initrd as well.

It is not necessary recommended for distributions to follow this
test case, in fact rescue (non-hostonly) initrd is not enabled by
default, but this is a very convinient way on the CI to test
both configuration and avoid regressions going forward.