ci: switch integration tests from ubuntu:latest to ubuntu:rolling
ubuntu:latest points to Ubuntu 24.04.

ubuntu:rolling currently points ot 24.10 and soon will point to
25.04.

This change has no impact on the "Daily Integration tests", which
anyways expect to test and pass both with ubuntu:latest and
ubuntu:rolling.

For Fedora fedora:latest points to the bi-yearly releases,
and the corresponding label for Ubuntu is ubuntu:rolling.