fix(release): tagging and release generation is no longer automated
Create release commit on relese branch automatically, but leave the
rest of the release creation process up to the project admins to
allow manual review and intervention.

We do not have a valid email address for the release automation,
so let's make it obvious that it is just a placeholder email address.